[Diagnostics][NFC] Move fix-it formatting into DiagnosticConsumers
Defer fix-it text formatting to individual DiagnosticConsumers so
it can be customized or omitted as needed, similar to how it's done
for errors, warnings, and notes

Move DiagnosticArgument, DiagnosticFormatOptions, and
formatDiagnosticText out of DiagnosticEngine and into
DiagnosticFormatting, to avoid circular dependencies between
DiagnosticEngine and DiagnosticConsumer
