+++ tools/SourceKit/lib/SwiftLang/SwiftEditor.cpp
@@ -108,7 +108,7 @@ void EditorDiagConsumer::handleDiagnostic(
   llvm::SmallString<256> Text;
   {
     llvm::raw_svector_ostream Out(Text);
-    DiagnosticEngine::formatDiagnosticText(Out, FormatString, FormatArgs);
+    DiagnosticFormatting::formatDiagnosticText(Out, FormatString, FormatArgs);
   }
   SKInfo.Description = Text.str();
