+++ tools/swift-api-digester/ModuleDiagsConsumer.cpp
@@ -108,7 +108,7 @@ void swift::ide::api::ModuleDifferDiagsConsumer::handleDiagnostic(
   llvm::SmallString<256> Text;
   {
     llvm::raw_svector_ostream Out(Text);
-    DiagnosticEngine::formatDiagnosticText(Out, FormatString, FormatArgs);
+    DiagnosticFormatting::formatDiagnosticText(Out, FormatString, FormatArgs);
   }
   AllDiags[Category].insert(Text.str().str());
 }
